refactor: combine model.py into core.py   (primarily due to inter-class dependencies in Article convenience methods such as Article.metrics)
